dealgood: remove gui
